requirements.txt
-e git+https://github.roche.com/BEDA/metadata_registration_dynamic_form.git#egg=DynamicForm
-e git+https://github.roche.com/BEDA/metadata_registration_state_machine.git#egg=StateMachine
Flask==1.1.1
flask-restx==0.1.0
Flask-WTF==0.14.2
gunicorn==20.0.4
json-logging-py==0.2
mongoengine==0.19.1
PyJWT==1.7.1
pymongo==3.10.1
PyYAML==5.3
requests==2.22.0
Werkzeug==0.16.0
WTForms==2.2.1
